+++ hive-event-listeners/apiary-gluesync-listener/src/main/java/com/expediagroup/apiary/extensions/gluesync/listener/service/HiveToGlueTransformer.java
@@ -1,3 +1,18 @@
+/**
+ * Copyright (C) 2019-2025 Expedia, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.expediagroup.apiary.extensions.gluesync.listener.service;
 
 import java.util.ArrayList;
@@ -6,6 +21,7 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 import org.apache.hadoop.hive.metastore.api.Database;
 import org.apache.hadoop.hive.metastore.api.FieldSchema;
@@ -26,6 +42,8 @@ public class HiveToGlueTransformer {
   private static final Logger log = LoggerFactory.getLogger(HiveToGlueTransformer.class);
   public static final String MANAGED_BY_GLUESYNC_KEY = "managed-by";
   public static final String MANAGED_BY_GLUESYNC_VALUE = "apiary-glue-sync";
+
+  private final GlueMetadataStringCleaner stringCleaner = new GlueMetadataStringCleaner();
 
   private final String gluePrefix;
 
@@ -74,7 +92,7 @@ public class HiveToGlueTransformer {
         .withStoredAsSubDirectories(storageDescriptor.isStoredAsSubDirectories());
 
     return new TableInput()
-        .withName(table.getTableName())
+        .withName(stringCleaner.clean(table.getTableName()))
         .withLastAccessTime(date)
         .withOwner(table.getOwner())
         .withParameters(table.getParameters())
@@ -91,7 +109,7 @@ public class HiveToGlueTransformer {
     final Collection<Column> columns = extractColumns(storageDescriptor.getCols());
 
     final SerDeInfo glueSerde = new SerDeInfo()
-        .withName(storageDescriptor.getSerdeInfo().getName())
+        .withName(stringCleaner.clean(storageDescriptor.getSerdeInfo().getName()))
         .withParameters(storageDescriptor.getSerdeInfo().getParameters())
         .withSerializationLibrary(storageDescriptor.getSerdeInfo().getSerializationLib());
 
@@ -110,11 +128,13 @@ public class HiveToGlueTransformer {
         .withSortColumns(sortOrders)
         .withStoredAsSubDirectories(storageDescriptor.isStoredAsSubDirectories());
 
+    List<String> partitionValues = partition.getValues().stream().map(stringCleaner::clean).collect(Collectors.toList());
+
     return new PartitionInput()
         .withLastAccessTime(date)
         .withParameters(partition.getParameters())
         .withStorageDescriptor(sd)
-        .withValues(partition.getValues());
+        .withValues(partitionValues);
   }
 
   public String glueDbName(String dbName) {
@@ -146,9 +166,9 @@ public class HiveToGlueTransformer {
 
     for (final FieldSchema fieldSchema : colList) {
       final Column col = new Column()
-          .withName(fieldSchema.getName())
-          .withType(fieldSchema.getType())
-          .withComment(fieldSchema.getComment());
+          .withName(stringCleaner.clean(fieldSchema.getName()))
+          .withType(stringCleaner.clean(fieldSchema.getType()))
+          .withComment(stringCleaner.shortTo254Chars(stringCleaner.clean(fieldSchema.getComment())));
 
       columns.add(col);
     }